third_party/WebKit/LayoutTests/fast/mediastream/argument-types-expected.txt

Tests the acceptable types for arguments to navigator.getUserMedia methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.webkitGetUserMedia() threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 0 present..
PASS navigator.webkitGetUserMedia(undefined) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia(null) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia({ }) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia({video: true}) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia(true) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia(42) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia(Infinity) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia(-Infinity) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia(emptyFunction) threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': 2 arguments required, but only 1 present..
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction) did not throw exception.
PASS navigator.webkitGetUserMedia(undefined, emptyFunction) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS navigator.webkitGetUserMedia(null, emptyFunction) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS navigator.webkitGetUserMedia({ }, emptyFunction) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS navigator.webkitGetUserMedia(true, emptyFunction) threw exception TypeError: Not an object..
PASS navigator.webkitGetUserMedia(42, emptyFunction) threw exception TypeError: Not an object..
PASS navigator.webkitGetUserMedia(Infinity, emptyFunction) threw exception TypeError: Not an object..
PASS navigator.webkitGetUserMedia(-Infinity, emptyFunction) threw exception TypeError: Not an object..
PASS navigator.webkitGetUserMedia(emptyFunction, emptyFunction) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS navigator.webkitGetUserMedia({video: true}, "foobar") threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, undefined) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, null) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, {}) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, true) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, 42) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, Infinity) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, -Infinity) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({ }, emptyFunction, emptyFunction) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, emptyFunction) did not throw exception.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, undefined) did not throw exception.
PASS navigator.webkitGetUserMedia({audio:true, video:true}, emptyFunction, undefined) did not throw exception.
PASS navigator.webkitGetUserMedia({audio:true}, emptyFunction, undefined) did not throw exception.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, "video") threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, null) did not throw exception.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, {}) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, true) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, 42) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, Infinity) threw exception TypeError: Type error.
PASS navigator.webkitGetUserMedia({video: true}, emptyFunction, -Infinity) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
